+++ packages/vue/src/components/molecules/SfSelect/SfSelect.vue
@@ -3,8 +3,8 @@
     v-click-outside="closeHandler"
     :aria-expanded="open.toString()"
     :aria-owns="'lbox_' + _uid"
-    aria-autocomplete="none"
-    role="combobox"
+    :aria-label="label"
+    role="listbox"
     :class="{
       'sf-select--is-active': isActive,
       'sf-select--is-selected': isSelected,
@@ -22,10 +22,9 @@
     <div style="position: relative;">
       <!-- eslint-disable-next-line vue/no-v-html -->
       <div
-        id="sfSelect"
+        ref="sfSelect"
         v-focus
         tabindex="0"
-        role="listbox"
         class="sf-select__selected sf-select-option"
         v-html="html"
       ></div>
@@ -39,7 +38,7 @@
       </slot>
       <SfOverlay :visible="open" class="sf-select__overlay mobile-only" />
       <transition name="sf-select">
-        <div v-show="open" role="list" class="sf-select__dropdown">
+        <div v-show="open" class="sf-select__dropdown">
           <!--  sf-select__option -->
           <ul
             :aria-expanded="open.toString()"
@@ -198,17 +197,15 @@ export default {
   mounted: function () {
     const options = [];
     const indexes = {};
-    let i = 0;
     if (!this.$slots.default) return;
     this.$on("update", this.update);
-    this.$slots.default.forEach((slot) => {
+    this.$slots.default.forEach((slot, index) => {
       if (!slot.tag) return;
       options.push({
         ...slot.componentOptions.propsData,
         html: slot.elm.innerHTML,
       });
-      indexes[JSON.stringify(slot.componentOptions.propsData.value)] = i;
-      i++;
+      indexes[JSON.stringify(slot.componentOptions.propsData.value)] = index;
     });
     this.options = options;
     this.indexes = indexes;
@@ -227,7 +224,7 @@ export default {
       if (index < 0) index = 0;
       if (index >= optionsLength) index = optionsLength - 1;
       this.index = index;
-      document.getElementById("sfSelect").blur();
+      this.$refs.sfSelect.blur();
       document.getElementById(this.focusedOption).focus();
     },
     enter() {